+++ obj/src/persistence/IdentifiableSqlServerPersistence.js
@@ -100,6 +100,10 @@ class IdentifiableSqlServerPersistence extends SqlServerPersistence_1.SqlServerP
      */
     constructor(tableName, schemaName) {
         super(tableName, schemaName);
+        /**
+         * Flag to turn on automated string ID generation
+         */
+        this._autoGenerateId = true;
     }
     /**
      * Converts the given object from the public partial format.
@@ -182,7 +186,7 @@ class IdentifiableSqlServerPersistence extends SqlServerPersistence_1.SqlServerP
         }
         // Assign unique id
         let newItem = item;
-        if (newItem.id == null) {
+        if (newItem.id == null && this._autoGenerateId) {
             newItem = Object.assign({}, newItem);
             newItem.id = item.id || pip_services3_commons_nodex_1.IdGenerator.nextLong();
         }
@@ -202,7 +206,7 @@ class IdentifiableSqlServerPersistence extends SqlServerPersistence_1.SqlServerP
                 return null;
             }
             // Assign unique id
-            if (item.id == null) {
+            if (item.id == null && this._autoGenerateId) {
                 item = Object.assign({}, item);
                 item.id = pip_services3_commons_nodex_1.IdGenerator.nextLong();
             }